Convert all integer primary keys to bigint

+++ core/store/models/sync_events.go
@@ -5,7 +5,7 @@ import "time"
 // SyncEvent represents an event sourcing style event, which is used to sync
 // data upstream with another service
 type SyncEvent struct {
-	ID        uint `gorm:"primary_key"`
+	ID        int64 `gorm:"primary_key"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	Body      string
